LayoutTests/platform/glib/imported/w3c/web-platform-tests/cookies/samesite/sandbox-iframe-nested.https-expected.txt



FAIL SameSite cookies with intervening sandboxed iframe and navigation assert_equals: `samesite_lax=0.35549777691779794` in request to `https://web-platform.test:9443`. expected (undefined) undefined but got (string) "0.35549777691779794"
